Added clown outpost.
Added the clown outpost to the adventure z-level. I also changed the
docking port for the clown's escape pod from centcom to a hanger area in
the clown outpost, meaning that any nerds that escape on it, I.E.
traitors, won't get credit for escaping to centcom. I might revert this
if people complain enough.